feat: Implement user authentication and enhance screens with SafeAreaView
- Added LoginScreen for user authentication with email and password.
- Integrated user role management and profile display in SettingsScreen.
- Updated HistoryScreen to fetch user-specific analysis history.
- Refactored screens (DarkHomeScreen, HomeScreen, NewHomeScreen, SettingsScreen) to use SafeAreaView for better layout handling.
- Enhanced API service to support user-specific analysis history retrieval.
- Introduced authService for handling login, registration, and user management.
- Updated types for user roles and analysis results to include image URLs.
- Added requirements.txt for backend dependencies.

+++ backend/app/models/analysis.py
@@ -28,7 +28,7 @@ class MLResults(BaseModel):
     """Machine learning prediction results"""
     prediction: PredictionResult
     confidence: float  # 0.0 - 1.0
-    processingTime: int  # in milliseconds
+    processingTime: float  # in seconds (can be decimal)
     modelVersion: str = "gwo-cnn-v1.0"
     features: List[float] = []  # GWO selected features
     rawOutput: float  # Raw model output before threshold
